requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
tika==2.6.0
python-dotenv==1.0.0
